Change LED subsystem to use ENUMs to auto update

+++ src/main/java/frc/robot/commands/intakerelease/ReleaseCMD.java
@@ -5,6 +5,7 @@ import frc.robot.subsystems.arm.Arm;
 import frc.robot.subsystems.intakerelease.IntakeRelease;
 import frc.robot.subsystems.led.Led2023;
 import frc.robot.subsystems.led.Led2023.COLORS_467;
+import frc.robot.subsystems.led.Led2023.ColorScheme;
 
 public class ReleaseCMD extends CommandBase {
   private final IntakeRelease intakerelease;
@@ -20,23 +21,24 @@ public class ReleaseCMD extends CommandBase {
   }
 
   @Override
+  public void initialize() {
+    ledStrip.set(COLORS_467.Black);
+  }
+
+  @Override
   public void execute() {
     if ((!arm.hasDropped())
         && intakerelease.haveCone()
-        && (
-        /*arm.getExtention() >= 0.18 &&*/ arm.getRotation() >= 0.1)) {
+        && (arm.getExtention() >= 0.18 && arm.getRotation() >= 0.1)) {
       arm.drop();
       return;
     }
     if (intakerelease.haveCube()) {
-      ledStrip.setColorMovingDown(
-          COLORS_467.Black.getColor(), COLORS_467.Purple.getColor()); // Purple,black
+      ledStrip.setCmdColorScheme(ColorScheme.RELEASE_CUBE);
     } else if (intakerelease.haveCone()) {
-
-      ledStrip.setColorMovingDown(
-          COLORS_467.Black.getColor(), COLORS_467.Yellow.getColor()); // Gold, black
+      ledStrip.setCmdColorScheme(ColorScheme.RELEASE_CONE);
     } else {
-      ledStrip.setColorMovingDownTwoClr(COLORS_467.Yellow.getColor(), COLORS_467.Purple.getColor());
+      ledStrip.setCmdColorScheme(ColorScheme.RELEASE_UNKNOWN);
     }
     intakerelease.release();
   }
